+++ lib/ui/page/home/tab/chats/view.dart
@@ -238,8 +238,7 @@ class ChatsTabView extends StatelessWidget {
                                     key: const Key('ArrowBack'),
                                     Icons.arrow_back_ios_new,
                                     size: 20,
-                                    color:
-                                        Theme.of(context).colorScheme.primary,
+                                    color: style.colors.primary,
                                   )
                                 : SvgImage.asset(
                                     'assets/icons/search.svg',
@@ -338,7 +337,7 @@ class ChatsTabView extends StatelessWidget {
                                 height: double.infinity,
                                 child: Icon(
                                   Icons.more_vert,
-                                  color: Theme.of(context).colorScheme.primary,
+                                  color: style.colors.primary,
                                 ),
                               ),
                             ),